refactor: isolate team loop commands
Motivation:
- Reduce complexity in the team orchestration module by moving pure loop command parsing out of AgentTeam.
- Make Ralph Wiggum loop command behavior easier to test and maintain independently.

Technical Changes:
- Added app.agent.mode.team.loop_commands for loop state, parsing, status payloads, and command detection.
- Updated team orchestration and team chat routes to import loop helpers from the new focused module.
- Added focused coverage for loop status payload progress calculation.
- Updated slash command documentation to point maintainers to the loop command parser.

Impact:
- Team orchestration is smaller and focused more on runtime coordination.
- Loop command behavior has a clearer ownership boundary and direct tests.
